moved files to one directory
